+++ quanmiaolightapp-20240801/ChangeLog.md
@@ -1,3 +1,15 @@
+2026-05-06 Version: 2.9.10
+- Update API GetEssayCorrectionTask: add response parameters Body.data.results.$.dimensionResults.
+- Update API GetEssayCorrectionTask: add response parameters Body.data.results.$.errorCode.
+- Update API GetEssayCorrectionTask: add response parameters Body.data.results.$.errorMessage.
+- Update API GetEssayCorrectionTask: add response parameters Body.data.results.$.overallComment.
+- Update API RunEssayCorrection: add request parameters dimensions.
+- Update API RunEssayCorrection: add response parameters Body.payload.output.dimensionResults.
+- Update API RunEssayCorrection: add response parameters Body.payload.output.overallComment.
+- Update API RunEssayCorrection: add response parameters Body.payload.output.reasoningContent.
+- Update API SubmitEssayCorrectionTask: add request parameters dimensions.
+
+
 2026-03-18 Version: 2.9.9
 - Generated php 2024-08-01 for QuanMiaoLightApp.